Reorganized code into individual folders.
